added models and routes folder
