+++ apis/app/models/project.py
@@ -1,0 +1,14 @@
+from pydantic import BaseModel
+from typing import Optional
+
+class ProjectCreate(BaseModel):
+    title: str
+    industry: str
+    objective: str
+
+class Project(ProjectCreate):
+    id: str
+    last_accessed: Optional[str] = None
+
+class ProjectUpdate(ProjectCreate):
+    id: str
