+++ apis/conf/.env.local
@@ -1,3 +1,3 @@
 NEO4J_URI=bolt://localhost:7687
 NEO4J_USER=neo4j
-NEO4J_PASSWORD=your_secure_password
+NEO4J_PASSWORD=testpassword
